requirements.txt
# CLI & Scaling
click==8.1.7
joblib==1.4.2

# Data Handling & Output
pandas==2.2.2
tabulate==0.9.0

# PDF Processing (text + scanned)
PyMuPDF==1.26.4   # modern, fast PDF text & image extractor
pytesseract==0.3.10  # OCR for scanned PDFs/images
Pillow==10.3.0       # image processing support for OCR











